move db interaction into separate module
