+++ stats/stats/src/charts/counters/average_block_time.rs
@@ -1,5 +1,5 @@
 use crate::{
-    charts::{
+    charts::db_interaction::{
         insert::{DateValue, DateValueDouble},
         updater::ChartFullUpdater,
     },
