+++ stats/stats/src/charts/db_interaction/insert.rs
