+++ stats/stats/src/charts/db_interaction/mod.rs
@@ -1,0 +1,4 @@
+//! Abstracted interaction with DB
+
+pub mod insert;
+pub mod updater;
