+++ stats/stats/src/charts/db_interaction/updater/full.rs
@@ -1,12 +1,9 @@
-//! Only retrieves new values and updates the latest one.
-//!
-//! In some cases performes full update (i.e. when some inconsistency was found or `force_full` is set)
+//! Re-reads/re-calculates whole chart from the source DB.
 
-use super::{get_last_row, get_min_block_blockscout};
 use crate::{
     charts::{
+        db_interaction::insert::{insert_data_many, DateValue},
         find_chart,
-        insert::{insert_data_many, DateValue},
     },
     metrics, Chart, UpdateError,
 };
@@ -14,42 +11,34 @@ use async_trait::async_trait;
 use sea_orm::prelude::*;
 
 #[async_trait]
-pub trait ChartPartialUpdater: Chart {
+pub trait ChartFullUpdater: Chart {
     async fn get_values(
         &self,
         blockscout: &DatabaseConnection,
-        last_row: Option<DateValue>,
     ) -> Result<Vec<DateValue>, UpdateError>;
 
     async fn update_with_values(
         &self,
         db: &DatabaseConnection,
         blockscout: &DatabaseConnection,
-        force_full: bool,
+        _force_full: bool,
     ) -> Result<(), UpdateError> {
         let chart_id = find_chart(db, self.name())
             .await
             .map_err(UpdateError::StatsDB)?
             .ok_or_else(|| UpdateError::NotFound(self.name().into()))?;
-        let min_blockscout_block = get_min_block_blockscout(blockscout)
-            .await
-            .map_err(UpdateError::BlockscoutDB)?;
-        // set offset to 1 because actual last row can be partially calculated
-        let offset = Some(1);
-        let last_row =
-            get_last_row(self, chart_id, min_blockscout_block, db, force_full, offset).await?;
         let values = {
             let _timer = metrics::CHART_FETCH_NEW_DATA_TIME
                 .with_label_values(&[self.name()])
                 .start_timer();
-            self.get_values(blockscout, last_row)
+            self.get_values(blockscout)
                 .await?
                 .into_iter()
-                .map(|value| value.active_model(chart_id, Some(min_blockscout_block)))
+                .map(|value| value.active_model(chart_id, None))
         };
         insert_data_many(db, values)
             .await
-            .map_err(UpdateError::StatsDB)?;
+            .map_err(UpdateError::BlockscoutDB)?;
         Ok(())
     }
 }
