+++ stats/stats/src/charts/lines/active_accounts.rs
@@ -1,5 +1,5 @@
 use crate::{
-    charts::{insert::DateValue, updater::ChartPartialUpdater},
+    charts::db_interaction::{insert::DateValue, updater::ChartPartialUpdater},
     UpdateError,
 };
 use async_trait::async_trait;
