+++ stats/stats/src/charts/lines/mock.rs
@@ -1,5 +1,5 @@
 use crate::{
-    charts::{insert::DateValue, updater::ChartFullUpdater},
+    charts::{db_interaction::insert::DateValue, db_interaction::updater::ChartFullUpdater},
     UpdateError,
 };
 use async_trait::async_trait;
