+++ stats/stats/src/charts/lines/native_coin_supply.rs
@@ -1,5 +1,5 @@
 use crate::{
-    charts::{
+    charts::db_interaction::{
         insert::{DateValue, DateValueDouble},
         updater::ChartPartialUpdater,
     },
