+++ stats/stats/src/charts/lines/new_accounts.rs
@@ -1,8 +1,10 @@
 use crate::{
     charts::{
         cache::Cache,
-        insert::{DateValue, DateValueInt},
-        updater::ChartFullUpdater,
+        db_interaction::{
+            insert::{DateValue, DateValueInt},
+            updater::ChartFullUpdater,
+        },
     },
     UpdateError,
 };
