+++ stats/stats/src/charts/lines/new_blocks.rs
@@ -1,5 +1,5 @@
 use crate::{
-    charts::{insert::DateValue, updater::ChartPartialUpdater},
+    charts::db_interaction::{insert::DateValue, updater::ChartPartialUpdater},
     UpdateError,
 };
 use async_trait::async_trait;
@@ -75,7 +75,7 @@ impl crate::Chart for NewBlocks {
 mod tests {
     use super::*;
     use crate::{
-        charts::updater::get_min_block_blockscout,
+        charts::db_interaction::updater::get_min_block_blockscout,
         get_chart_data,
         tests::{init_db::init_db_all, mock_blockscout::fill_mock_blockscout_data},
         Chart,
