+++ stats/stats/src/charts/mod.rs
@@ -1,9 +1,7 @@
 pub mod cache;
 mod chart;
 pub mod counters;
-pub mod insert;
+pub mod db_interaction;
 pub mod lines;
 mod mutex;
-pub mod updater;
-
 pub use chart::{create_chart, find_chart, Chart, MissingDatePolicy, UpdateError};
